requirements.txt
python-xlib==0.33
six==1.16.0
